Accommodate restrictions of released stl-rs
This mainly just means we have to clone a little more and cast things
between f32 and f64, because the released version of stl-rs doesn't
allow us to take ownership of the various components and doesn't include
f64 compatibility.

Note that this still currently points towards the main git branch for
stl-rust, but at least all the incorporated changes are more likely to
be merged sooner or later!

Unfortunately this results in some benchmark regressions, probably due
to the extra clones and having to convert to/from f32 :(

```
     Running benches/vic_elec.rs (target/release/deps/vic_elec-36a0fad3091799a8)
vic_elec                time:   [28.569 ms 28.591 ms 28.619 ms]
                        change: [+11.550% +11.750% +11.918%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

```