Use stlrs MSTL implementation
stlrs will have its own implementation of MSTL in v0.3.0 which means
less for this crate to maintain. Benchmarks show it runs pretty much
identically to the augurs implementation which is nice. The only
downside is the f64 -> f32 conversion required before calling the stlrs
APIs but that would be required anyway if we want to avoid any forks.