Use `tsify` to avoid `JsValue`/`any` types
Prior to this commit, the JS binding functions returned untyped `JsValue`s,
which meant the generated TypeScript types for the functions had `any`
as the return value.

This commit uses the `tsify` crate to derive the correct traits for
types returned from WASM, so that our functions can return those types
and the TypeScript definitions reflect that.