Use PreprocessedData associated type in outlier trait
This lets us preprocess the data in one step and use it for
multiple detection passes with different hyperparameters without
needing to preprocess it again every time.

The tradeoff is that the trait object is no longer trivially boxable
since the associated type will need to be defined. We can reconsider
this in a future version if it's important.