fix: use more accurate series count to avoid panic
Prior to this commit we used the number of values at the _first_ timestamp, _after_ omitting NANs, to determine the total number of series. This is inaccurate if there are any missing values at that timestamp, which caused a panic later when we tried to index into the series vector for a later timestamp.

Instead we now keep track of the number of series as we sort the data by tracking the max of all timestamps, and use this when creating the vec of series instead.