feat: add cmdstan-based optimizer for augurs-prophet
This commit adds an implementation of `Optimizer` to the
`augurs-prophet` crate that uses `cmdstan` to optimize or sample
from the Prophet model, similar to the Prophet Python package.

The `cmdstan` feature is disabled by default since it requires external
dependencies and may not work well on all platforms.

The `compile-cmdstan` feature can be enabled to build the Prophet model
at build time. This requires a working installation of `stan` and setting
the `STAN_PATH` environment variable to the path to the `stan` installation.
This will embed the Prophet model binary and the libtbb dynamic library
into the final executable, which will increase the size of the final
executable by about 2MB, but the final binary won't require any additional
dependencies.

There's an example in the `examples` directory which fits and predicts
using cmdstan. It can be runwith something like:

```bash
STAN_PATH=/home/ben/.conda/envs/stan cargo run --example prophet_cmdstan
```

(replacing the `STAN_PATH` value with the path to your Stan installation).