feat: add Prophet functionality to augurs-js
This adds Javascript bindings for the augurs-prophet crate
as part of the augurs Javascript library.

When creating the Prophet object, an optimizer must be
provided by the user. This is to avoid having a hard dependency
on the Stan WASM module which increases the size of the augurs
WASM blob significantly and won't be needed for users not using
Prophet.