feat!: split JS package into separate crates
This enables us to package up each piece of functionality as a separate
WASM module, which means users can import only the functionality they
need to save on bundle size.

It doesn't work as-is because some types are expected to work across
packages but they're technically considered distinct types after
compilation. It's possible that doing this using WASM components and
JCO will achieve this but I'm not certain.