Unexpose DistanceMatrix from JS; convert in each JS crate instead
Before this commit we were using a shared DistanceMatrix type from augurs-core-js,
and assuming that we could return it from one WASM module and use it directly in
a second. This doesn't work, however; those are separate types at runtime and the
memory is not shared between the two WASM modules.

Instead, just return an array of Float64Arrays when calculating the distance
matrix and accept that in the clustering side.