feat: add Forecaster wrapper for Prophet
Since the `Forecaster` traits don't pass timestamps etc,
we need to provide the training data at the time the forecaster
is created and pass _that_ down to the Prophet model, after replacing
the `y` column with whatever the forecaster gives us. This should
work with transforms too.

It does require cloning the data, unfortunately. The alternative
approach which I tried first was to use an associated type for the
data, but that is pretty infectious and ends up requiring the input
data to be mutable so that we can replace the `y` column, which ends
up looking pretty ugly.