fix: correctly set holiday feature to 1 for all holiday dates
Prior to this commit, the holiday feature was only set to 1 for the
first occurrence of a holiday date. This meant that the holiday
effect was only applied to predictions for that single datetime
in the dataframe, rather than all dates in the dataframe that
were on the holidays.

This commit fixes that by iterating over all dates in the dataframe
that are on the holiday, and setting the holiday feature to 1.0 for
those dates.