feat(prophet): support sub-daily & non-UTC holidays
Prior to this commit, there was an implicit assumption that all holidays
lasted one day, and that day lasted from midnight to midnight
UTC. This partly stemmed from how the Python Prophet implementation
works (although that supports alternative timezones by using non-UTC
timestamp values).

This commit changes the Holidays APIs to require each holiday to be
specified using a start and end Unix timestamp. Whether a datapoint is
considered 'on holiday' is then a simple check of whether its timestamp
lies within the holiday start/end times. This allows us to support:

- non-UTC holidays
- sub-daily holidays
- multi-day holidays

since the user can provide whatever timestamps they want for the
holiday.

One thing we lose here is the lower and upper windows for holidays,
which originally (in the Python implementation at least) added a further
feature to the Stan model for each day encompassed by the window. This
can be worked around manually by just adding separate holiday
occurrences for those windows. The reason they are no longer supported
is because now that holidays are no longer a fixed width (1 day) it's
not obvious how wide each 'window' should be.