fix: add a separate feature for each holiday's lower/upper windows
Reading through the original Prophet Python code, it looks like we're
supposed to add a single feature for each value in the window of each
holiday, rather than a separate feature every time the holiday occurs.
This commit does so by creating a map of features and updating the
relevant column in the map while iterating over the holidays
occurrences.