fix: make Transform enum non-exhaustive
Happily, cargo-semver-checks caught the addition of some variants
in #185, but it's a shame that had to be a minor bump. This marks
the enum as non-exhaustive so future transform additions won't
necessarily be considered breaking changes. It's unlikely users
will ever have to match on this anyway so this shouldn't cause
any problems (it could probably be a struct with an 'inner'
enum, really).