feat: add 'transforms' JS crate and include in augurs JS bindings
This provides a way to access the power transform functions from JS.

I'd like to add the Forecaster functionality to the Prophet bindings
somehow but that limits the Prophet APIs to only return yhat and
the bounds. That's probably enough for many cases but it's not ideal.

This is a quick workaround to get the transform functionality into the
JS package really.