feat(js): standardize data as part of power transformation (#205)
* feat: add standard scaler transform

This PR adds a standard scaler transform to the augurs-forecaster
transforms. It's similar to the existing `MinMaxScaler`, but instead
scales the data such that it has a mean of 0 and a standard deviation
of 1. This is similar to the `StandardScaler` in scikit-learn.

* feat(js): standardize data as part of power transformation

This is the default in scikit-learn's PowerTransformer and helps deal with
extreme values in the input.