Commit prophet-wasmstan.wasm to git
Until now we've avoided committing the prophet-wasmstan.wasm file to git.
This is because it's a binary file and we don't want to bloat the repo.
However, it's annoying for contributors to have to build the WASM
themselves, since it requires lots of undocumented dependencies. It also
slows down CI considerably since we need to build the WASM every time
we want to run any kind of tests.

This commit adds the WASM file to the repo in the final expected
location (crates/augurs-prophet/prophet-wasmstan.wasm), removes the
WASM build steps as dependencies from tests in CI, and modifies the
existing wasmstan CI pipeline to ensure that the built WASM file
matches the one in the repo. This avoids the need for the WASM build
to happen before the tests, and also ensures that the WASM file is
always up to date. It also means that contributors can run tests
without having to build the WASM themselves.