Scale the data before or after a transformation
Allow users to specify if they do not want to scale their data, or to
scale it either before or after doing a power transformation. This
allows both matching the sklearn behavior of scaling the data after the
transformation, or scaling it before the transformation which can help
with data that floors at non-zero values.