refactor(clustering)!: use new DbscanCluster type instead of isize
Add a new `DbscanCluster` type to represent the cluster labels, which
can be either a cluster ID or a special value indicating noise. The
`Cluster` variant is now a `NonZeroU32` instead of an `isize`.

We could use a `u32` but using a `NonZeroU32` allows us to make use
of the niche optimizations that `NonZeroU32` provides and halves the
size of the type. Questionable whether it's worth it though?

Closes #229.