feat!: allow ignoring NaNs in JS transforms, too
Unfortunately this requires a breaking change to the JS API;
while transforms were previously specified as an array of strings,
they are now specified as an array of objects with a `type` property.

This is a breaking change, but it's a sensible one, because it will
allow us to add further options to transforms in future.

BREAKING CHANGE: transforms are now specified as an array of objects
with a `type` property.