feat: add Wasm components for outlier and changepoint detection; generate Go bindings
This is a proof of concept of using [`gravity`] to generate Go bindings
for some of augurs' functionality. There are two stages:

1. Build Wasm components which wraps some of `augurs` key methods
2. Generate Go bindings from the Wasm component using gravity.

This results in a Go package with some subpackages that can be used to
call augurs code.

I'm using a fork of `gravity` for now to get a few more features, which
will hopefully be merged upstream soon.

[`gravity`]: https://github.com/sd2k/gravity