fix: assert that data is rectangular in DBSCAN preprocessing
Prior to this commit, the DBSCAN preprocessing step would panic with a
confusing error messages during transposition if the data was not
rectangular. This was hard to debug because it didn't mention the
expected shape of the data.

This commit adds a check that the data is rectangular and panics if it
is not. It also adds tests to ensure that the check is performed.

It also adds a fallible version of the preprocessing step that returns
an error if the data is not rectangular.

Relates to #362.